backend/requirements.txt
Django==5.2.4
djangorestframework==3.16.0
django-cors-headers==4.7.0
pillow==11.3.0
stripe==12.4.0
python-decouple==3.8
django-extensions==4.1
psycopg2-binary==2.9.10
celery==5.5.3
redis==6.2.0
channels==4.3.0
channels-redis==4.3.0
django-filter==25.1